fix: Refactor to use more consistent barrel composition for ts structure
