+++ src/admin/components/campaigns/index.ts
@@ -1,0 +1,4 @@
+export * from './tabs';
+export * from './CampaignEdit';
+export * from './CampaignsTable';
+export * from './GenerateShortcode';
