+++ src/admin/components/controls/index.ts
@@ -1,9 +1,10 @@
-export * from './TextControl';
-export * from './RadioGroupControl';
-export * from './RadioControl';
-export * from './ToggleControl';
-export * from './TextAreaControl';
-export * from './ColorPicker';
-export * from './SelectControl';
-export * from './FormTokenField';
+export * from './BaseControl';
 export * from './CheckboxControl';
+export * from './ColorPicker';
+export * from './FormTokenField';
+export * from './RadioControl';
+export * from './RadioGroupControl';
+export * from './SelectControl';
+export * from './TextAreaControl';
+export * from './TextControl';
+export * from './ToggleControl';
