+++ src/admin/components/donors/index.ts
@@ -1,0 +1,1 @@
+export * from './DonorsTable';
