+++ src/admin/components/settings/index.ts
@@ -1,0 +1,2 @@
+export * from './tabs';
+export * from './SettingsPage';
