+++ src/admin/components/table/index.ts
@@ -1,0 +1,5 @@
+export * from './DetailsModal';
+export * from './Pagination';
+export * from './Search';
+export * from './Table';
+export * from './TableControls';
