+++ src/admin/components/transactions/index.ts
@@ -1,0 +1,1 @@
+export * from './TransactionsTable';
