+++ src/admin/kudos-admin.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import domReady from '@wordpress/dom-ready';
 import './kudos-admin.css';
 import { AdminProvider, SettingsProvider } from './contexts';
-import { AdminRouter } from './components/AdminRouter';
+import { AdminRouter } from './components';
 
 domReady(() => {
 	const container = document.getElementById('root');
