+++ src/block/components/controls/index.ts
@@ -1,9 +1,8 @@
-export * from './ToggleControl';
-export * from './TextControl';
-export * from './TextAreaControl';
-export * from './Button';
-export * from './RadioGroupControl';
-export * from './CheckboxControl';
-export * from './TextAreaControl';
-export * from './SelectControl';
 export * from './BaseController';
+export * from './Button';
+export * from './CheckboxControl';
+export * from './RadioGroupControl';
+export * from './SelectControl';
+export * from './TextAreaControl';
+export * from './TextControl';
+export * from './ToggleControl';
