+++ src/block/components/index.ts
@@ -1,0 +1,12 @@
+export * from './tabs';
+export * from './controls';
+export * from './DonateButton';
+export * from './Edit';
+export * from './FormRouter';
+export * from './KudosForm';
+export * from './KudosLogo';
+export * from './KudosModal';
+export * from './Message';
+export * from './ProgressBar';
+export * from './Render';
+export * from './Spinner';
