+++ src/block/components/tabs/index.ts
@@ -1,4 +1,5 @@
 export * from './AddressTab';
+export * from './BaseTab';
 export * from './FrequencyTab';
 export * from './InitialTab';
 export * from './MessageTab';
