+++ src/types/images.d.ts
@@ -1,2 +1,5 @@
 declare module '*.svg';
-declare module '*.png';
+declare module '*.png' {
+	const value: string;
+	export default value;
+}
