+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"include": ["src", "types"],
+	"include": ["src", "src/types"],
 	"exclude": ["node_modules", "build"],
 	"compilerOptions": {
 		"resolveJsonModule": true,
